chore(quality): address #12 low-priority items — dead_code cleanup, log redaction, cargo audit
PR #5 引き継ぎ項目:
- 各 crate の crate-wide `#![allow(dead_code, unused_imports, unused_variables)]`
  を除去し、必要な箇所は struct / enum / impl レベルの個別 `#[allow(dead_code)]`
  (将来用途のため残す API に限る) に置換。
- `ProjectManager::open/close/list` の後方互換 thin wrapper を削除
  (trait `ProjectConfiguration` の `open_project/close_project/list_projects`
  と完全重複していた)。`close_all` は Daemon shutdown 経路で使用中なので残置。
- `cargo fix` で unused imports を自動除去。

ログ露出削減 (S24 一部):
- `PeerId::short()` を追加: 先頭 12 文字 + `…` の短縮表示。
- conduit/quic の `tracing::info!("... peer {} ...")` を `peer_id.short()` に
  差し替え (4 箇所)。全箇所への波及は次フェーズで。

CI 強化:
- `.github/workflows/ci.yml` に `cargo audit` job を追加
  (`--deny unsound --deny yanked` で unsound + yanked クレートをブロック。
  unmaintained 警告は当面許容)。

検証: cargo fmt --all / clippy -D warnings / test --all (34/34 pass)。

+++ synergos-net/src/quic/mod.rs
@@ -3,7 +3,6 @@
 //! quinn を使った QUIC 接続の確立・管理・ストリーム制御を提供する。
 //! サーバー（リスナー）とクライアント（コネクタ）の両方の役割を担う。
 
-use std::collections::HashMap;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use std::time::Duration;
@@ -178,7 +177,7 @@ impl QuicManager {
             .as_ref()
             .ok_or_else(|| SynergosNetError::Quic("Endpoint not initialized".into()))?;
 
-        tracing::info!("Connecting to peer {} at {}", peer_id, addr);
+        tracing::info!("Connecting to peer {} at {}", peer_id.short(), addr);
 
         let connection = endpoint
             .connect(addr, server_name)
@@ -212,7 +211,7 @@ impl QuicManager {
         }
         self.connections.remove(peer_id);
         self.bandwidth_estimates.remove(peer_id);
-        tracing::info!("Disconnected from peer {}: {}", peer_id, reason);
+        tracing::info!("Disconnected from peer {}: {}", peer_id.short(), reason);
     }
 
     /// 指定ピアとの双方向ストリームを開く
